+++ src/main/java/com/calibermc/buildify/datagen/properties/CraftingTableTypes.java
@@ -1,23 +1,25 @@
 package com.calibermc.buildify.datagen.properties;
 
 public enum CraftingTableTypes {
-    ANVIL("anvil"),
-    BEEHIVE("beehive"),
-    BLAST_FURNACE("blast_furnace"),
-    BREWING_STAND("brewing_stand"),
-    CARTOGRAPHY_TABLE("cartography_table"),
-    CAULDRON("cauldron"),
-    CRAFTING_TABLE("crafting_table"),
-    ENCHANTING_TABLE("enchanting_table"),
-    FLETCHING_TABLE("fletching_table"),
-    FURNACE("furnace"),
-    GRINDSTONE("grindstone"),
-    LECTERN("lectern"),
-    LOOM("loom"),
-    SMOKER("smoker"),
-    SMITHING_TABLE("smithing_table"),
-    STONECUTTER("stonecutter"),
-    WOODCUTTER("woodcutter");
+    ANVIL("minecraft:anvil"),
+    BEEHIVE("minecraft:beehive"),
+    BLAST_FURNACE("minecraft:blast_furnace"),
+    BREWING_STAND("minecraft:brewing_stand"),
+    CARTOGRAPHY_TABLE("minecraft:cartography_table"),
+    CAULDRON("minecraft:cauldron"),
+    CRAFTING_TABLE("minecraft:crafting_table"),
+    ENCHANTING_TABLE("minecraft:enchanting_table"),
+    FLETCHING_TABLE("minecraft:fletching_table"),
+    FURNACE("minecraft:furnace"),
+    GRINDSTONE("minecraft:grindstone"),
+    LECTERN("minecraft:lectern"),
+    LOOM("minecraft:loom"),
+    SMITHING_TABLE("minecraft:smithing_table"),
+    SMOKER("minecraft:smoker"),
+    STONECUTTER("minecraft:stonecutter"),
+
+    // Caliber
+    WOODCUTTER("caliber:woodcutter");
 
     private final String name;